refactor(#46): 클래스 생성 가격 페이지 컴포넌트 분할 및 수강생 옵션 함수 분리
